episode_100

+++ priv/repo/migrations/20200309041124_create_supuls.exs
@@ -4,7 +4,7 @@ defmodule Demo.Repo.Migrations.CreateSupuls do
   def change do
     create table(:supuls, primary_key: false) do
       add :id, :uuid, primary_key: true
-      add :supul_code, :decimal, precision: 8
+      add :supul_code, :integer
       add :name, :string
       add :geographical_area, :string
